Cleaned up the file system by removing old files and renaming others
