+++ frontend/src/App.tsx
@@ -2,8 +2,8 @@ import { Outlet, Route, Routes } from "react-router-dom";
 
 import { NavBar } from "./components/index.ts";
 import { Apply, Candidates } from "./pages/index.ts";
-import UploadButtonWithPopup from "./components/ButtonForFileUploadPopup";
-import TestButton from "./components/PopupWindow.tsx";
+
+import TestButton from "./components/FileUploadPopupWindow.tsx";
 
 function Layout() {
   return (
@@ -17,7 +17,6 @@ function Layout() {
 function App() {
   return (
     <div className="app-container">
-      <UploadButtonWithPopup />
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route index element={<Candidates />} />
