+++ frontend/src/components/FileUploadPopupWindow.tsx
