--- frontend/src/components/modalStyles.css
@@ -1,14 +1,0 @@
-.modal-content {
-  width: 536px;
-  height: 433px;
-  top: 339px;
-  left: 452px;
-  border-radius: 8px;
-  gap: 10px;
-  background-color: #ffffff;
-}
-
-.react-modal-overlay {
-  opacity: 0.7;
-  background-color: rgba(0, 0, 0, 0.7);
-}
